Combined two small files into one.
